requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
markdown2==2.5.0
bleach==6.1.0
python-slugify==8.0.4
python-dotenv==1.0.1
